fold GoogleAppLegacy into Gmail
Gmail was the only remaining subclass of GoogleAppLegacy after #541.
Move the base class body into Gmail directly, drop the inheritance,
and delete google-app-legacy.ts. Behavior is preserved verbatim.

The hooks.beforeLoadUrl extension point and webContentsViewOptions
constructor option existed only to let Gmail pass preload args and
CSS-injection listeners up to the base class; both are inlined into
Gmail.createView now that there is no base class.
