+++ packages/app/gmail/index.ts
@@ -1,22 +1,37 @@
 import fs from "node:fs";
 import path from "node:path";
-import { platform } from "@electron-toolkit/utils";
+import { is, platform } from "@electron-toolkit/utils";
+import { accountColorsMap } from "@meru/shared/accounts";
+import { APP_TITLEBAR_HEIGHT, GOOGLE_ACCOUNTS_URL } from "@meru/shared/constants";
 import {
   createGmailDelegatedAccountUrl,
+  GMAIL_DELEGATED_ACCOUNT_URL_REGEXP,
   GMAIL_INBOX_FEED_URL,
   GMAIL_PRELOAD_ARGUMENTS,
   GMAIL_URL,
   type GmailInboxMessage,
+  isGmailComposeWindowUrl,
 } from "@meru/shared/gmail";
 import { getGoogleAppUrl } from "@meru/shared/google";
-import type { GoogleAppsPinnedApp } from "@meru/shared/types";
-import { app, BrowserWindow, clipboard } from "electron";
+import { supportedGoogleApps } from "@meru/shared/types";
+import type { GoogleAppsPinnedApp, SupportedGoogleApp } from "@meru/shared/types";
+import {
+  app,
+  BrowserWindow,
+  clipboard,
+  dialog,
+  globalShortcut,
+  powerSaveBlocker,
+  type Session,
+  WebContentsView,
+  type WebContentsViewConstructorOptions,
+} from "electron";
 import { subscribeWithSelector } from "zustand/middleware";
 import { createStore } from "zustand/vanilla";
 import { accounts } from "@/accounts";
 import { config } from "@/config";
 import { setupWindowContextMenu } from "@/context-menu";
-import { GoogleAppLegacy, type GoogleAppLegacyOptions } from "@/google-app-legacy";
+import { GoogleApp } from "@/google-app";
 import { ipc } from "@/ipc";
 import { licenseKey } from "@/license-key";
 import { main } from "@/main";
@@ -30,12 +45,23 @@ import z from "zod";
 import { createNotification, isWithinNotificationTimes } from "@/notifications";
 import { ms } from "@meru/shared/ms";
 import { wait } from "@meru/shared/utils";
+import { openExternalUrl } from "@/url";
 
 export const GMAIL_USER_STYLES_PATH = path.join(app.getPath("userData"), "gmail-user-styles.css");
 
 const GMAIL_USER_STYLES: string | null = fs.existsSync(GMAIL_USER_STYLES_PATH)
   ? fs.readFileSync(GMAIL_USER_STYLES_PATH, "utf-8")
   : null;
+
+const WINDOW_OPEN_URL_WHITELIST = [
+  /googleusercontent\.com\/viewer\/secure\/pdf/, // Print PDF
+];
+
+const SUPPORTED_GOOGLE_APPS_URL_REGEXP = new RegExp(
+  `(${Object.keys(supportedGoogleApps).join("|")})(?:\\.usercontent)?\\.google\\.com`,
+);
+
+const WINDOW_OPEN_DOWNLOAD_URL_WHITELIST = [/chat\.google\.com\/u\/\d\/api\/get_attachment_url/];
 
 const inboxFeedEntryAuthorSchema = z.object({
   name: z.coerce.string(),
@@ -189,7 +215,47 @@ function extractVerificationCode(texts: string[]) {
   return null;
 }
 
-export class Gmail extends GoogleAppLegacy {
+export class Gmail {
+  accountId: string;
+
+  url: string;
+
+  baseUrl: string;
+
+  session: Session;
+
+  private additionalArguments: string[];
+
+  private _view: WebContentsView | undefined;
+
+  get view() {
+    if (!this._view) {
+      throw new Error("View has not been created yet");
+    }
+
+    return this._view;
+  }
+
+  set view(view: WebContentsView) {
+    this._view = view;
+  }
+
+  viewStore = createStore(
+    subscribeWithSelector<{
+      navigationHistory: {
+        canGoBack: boolean;
+        canGoForward: boolean;
+      };
+      attentionRequired: boolean;
+    }>(() => ({
+      navigationHistory: {
+        canGoBack: false,
+        canGoForward: false,
+      },
+      attentionRequired: false,
+    })),
+  );
+
   userEmail: string | null = null;
 
   unreadCountEnabled = true;
@@ -223,10 +289,12 @@ export class Gmail extends GoogleAppLegacy {
     unifiedInboxEnabled,
     delegatedAccountId,
   }: {
+    accountId: string;
+    session: Session;
     unreadCountEnabled: boolean;
     unifiedInboxEnabled: boolean;
     delegatedAccountId: string | null;
-  } & Omit<GoogleAppLegacyOptions, "url">) {
+  }) {
     const additionalArguments: string[] = [];
 
     if (config.get("gmail.hideGmailLogo")) {
@@ -267,42 +335,15 @@ export class Gmail extends GoogleAppLegacy {
       }
     }
 
-    super({
-      accountId,
-      url: delegatedAccountId ? createGmailDelegatedAccountUrl(delegatedAccountId) : GMAIL_URL,
-      session,
-      webContentsViewOptions: {
-        webPreferences: {
-          preload: getPreloadPath("gmail"),
-          additionalArguments,
-        },
-      },
-      hooks: {
-        beforeLoadUrl: [
-          (view) => {
-            view.webContents.on("dom-ready", () => {
-              if (view.webContents.getURL().startsWith(GMAIL_URL)) {
-                view.webContents.insertCSS(gmailCSS);
+    this.accountId = accountId;
 
-                if (licenseKey.isValid && GMAIL_USER_STYLES) {
-                  view.webContents.insertCSS(GMAIL_USER_STYLES);
-                }
-              }
+    this.url = delegatedAccountId ? createGmailDelegatedAccountUrl(delegatedAccountId) : GMAIL_URL;
 
-              view.webContents.insertCSS(meruCSS);
-            });
+    this.baseUrl = new URL(this.url).origin;
 
-            view.webContents.on("did-navigate-in-page", (_event, url) => {
-              const hash = new URL(url).hash;
+    this.session = session;
 
-              const messageIdMatch = hash.match(/#[^/]+\/([A-Za-z0-9]{15,})$/);
-
-              this.store.setState({ messageId: messageIdMatch?.[1] || null });
-            });
-          },
-        ],
-      },
-    });
+    this.additionalArguments = additionalArguments;
 
     this.unreadCountEnabled = unreadCountEnabled;
 
@@ -317,6 +358,385 @@ export class Gmail extends GoogleAppLegacy {
         }
       }
     }, ms("5m"));
+  }
+
+  createView(options?: WebContentsViewConstructorOptions) {
+    this.view = new WebContentsView({
+      ...options,
+      webPreferences: {
+        preload: getPreloadPath("gmail"),
+        additionalArguments: this.additionalArguments,
+        ...options?.webPreferences,
+        session: this.session,
+      },
+    });
+
+    main.window.contentView.addChildView(this.view);
+
+    this.registerNavigationHandler(this.view);
+
+    this.registerFoundInPageHandler();
+
+    this.registerWindowOpenHandler(this.view);
+
+    this.view.webContents.on("dom-ready", () => {
+      this.view.webContents.setVisualZoomLevelLimits(1, 3);
+    });
+
+    setupWindowContextMenu(this.view);
+
+    this.updateViewBounds();
+
+    main.window.on("resize", () => {
+      this.updateViewBounds();
+    });
+
+    this.view.webContents.on("dom-ready", () => {
+      if (this.view.webContents.getURL().startsWith(GMAIL_URL)) {
+        this.view.webContents.insertCSS(gmailCSS);
+
+        if (licenseKey.isValid && GMAIL_USER_STYLES) {
+          this.view.webContents.insertCSS(GMAIL_USER_STYLES);
+        }
+      }
+
+      this.view.webContents.insertCSS(meruCSS);
+    });
+
+    this.view.webContents.on("did-navigate-in-page", (_event, url) => {
+      const hash = new URL(url).hash;
+
+      const messageIdMatch = hash.match(/#[^/]+\/([A-Za-z0-9]{15,})$/);
+
+      this.store.setState({ messageId: messageIdMatch?.[1] || null });
+    });
+
+    if (is.dev) {
+      this.view.webContents.openDevTools({ mode: "bottom" });
+    }
+
+    return this.view.webContents.loadURL(this.url);
+  }
+
+  private registerNavigationHandler(window: BrowserWindow | WebContentsView) {
+    window.webContents.on("did-navigate", (_event, url) => {
+      if (url.startsWith(`${GOOGLE_ACCOUNTS_URL}/v3/signin/challenge/pk/presend`)) {
+        dialog.showMessageBox({
+          type: "info",
+          message: "Passkey sign-in not supported yet",
+          detail: "Please use password to sign in.",
+        });
+      }
+
+      if (window === this.view) {
+        this.viewStore.setState({
+          navigationHistory: {
+            canGoBack: this.view.webContents.navigationHistory.canGoBack(),
+            canGoForward: this.view.webContents.navigationHistory.canGoForward(),
+          },
+          attentionRequired: !url.startsWith(this.baseUrl),
+        });
+      }
+    });
+
+    if (window === this.view) {
+      window.webContents.on("did-navigate-in-page", (_event: Electron.Event) => {
+        this.viewStore.setState({
+          navigationHistory: {
+            canGoBack: this.view.webContents.navigationHistory.canGoBack(),
+            canGoForward: this.view.webContents.navigationHistory.canGoForward(),
+          },
+        });
+      });
+    }
+
+    window.webContents.on("will-redirect", (event, url) => {
+      if (
+        url.startsWith("https://www.google.com") ||
+        url.startsWith("https://workspace.google.com")
+      ) {
+        event.preventDefault();
+
+        window.webContents.loadURL(`${GOOGLE_ACCOUNTS_URL}/ServiceLogin?service=mail`);
+      }
+    });
+  }
+
+  private registerFoundInPageHandler() {
+    this.view.webContents.on("found-in-page", (_event, result) => {
+      ipc.renderer.send(main.window.webContents, "findInPage.result", {
+        activeMatch: result.activeMatchOrdinal,
+        totalMatches: result.matches,
+      });
+    });
+  }
+
+  updateViewBounds() {
+    const { width, height } = main.getWindowBounds();
+
+    this.view.setBounds({
+      x: 0,
+      y: APP_TITLEBAR_HEIGHT,
+      width,
+      height: height - APP_TITLEBAR_HEIGHT,
+    });
+  }
+
+  destroy() {
+    this.view.webContents.removeAllListeners();
+
+    this.view.webContents.close();
+
+    this.view.removeAllListeners();
+
+    main.window.contentView.removeChildView(this.view);
+  }
+
+  registerWindowOpenHandler(window: BrowserWindow | WebContentsView) {
+    window.webContents.setWindowOpenHandler(({ url, disposition }) => {
+      if (url === "about:blank") {
+        return {
+          action: "allow",
+          createWindow: (options) => {
+            let newWindow: BrowserWindow | null = new BrowserWindow({
+              ...options,
+              show: false,
+            });
+
+            newWindow.webContents.once("will-navigate", (_event, url) => {
+              if (newWindow) {
+                if (url.startsWith(GOOGLE_ACCOUNTS_URL)) {
+                  newWindow.show();
+
+                  return;
+                }
+
+                openExternalUrl(url);
+
+                newWindow.webContents.close();
+
+                newWindow = null;
+              }
+            });
+
+            return newWindow.webContents;
+          },
+        };
+      }
+
+      if (
+        (url.startsWith(GMAIL_URL) && url.includes("view=pt")) ||
+        url.startsWith(GOOGLE_ACCOUNTS_URL)
+      ) {
+        return {
+          action: "allow",
+        };
+      }
+
+      const matchedSupportedGoogleApp = url.match(SUPPORTED_GOOGLE_APPS_URL_REGEXP)?.[1] as
+        | SupportedGoogleApp
+        | undefined;
+
+      const isGoogleAppEnabledToOpenInApp =
+        licenseKey.isValid &&
+        matchedSupportedGoogleApp &&
+        config.get("googleApps.openInApp") &&
+        !config.get("googleApps.openInAppExcludedApps").includes(matchedSupportedGoogleApp);
+
+      if (
+        (url.startsWith(GMAIL_URL) ||
+          WINDOW_OPEN_URL_WHITELIST.some((regex) => regex.test(url)) ||
+          isGoogleAppEnabledToOpenInApp) &&
+        disposition !== "background-tab"
+      ) {
+        if (matchedSupportedGoogleApp) {
+          if (
+            !config.get("googleApps.openAppsInNewWindow") &&
+            GoogleApp.reuseWindowByHostname(this.accountId, url)
+          ) {
+            return { action: "deny" };
+          }
+
+          new GoogleApp({
+            accountId: this.accountId,
+            url,
+          });
+
+          return { action: "deny" };
+        }
+
+        const gmailDelegatedAccountId = url.match(GMAIL_DELEGATED_ACCOUNT_URL_REGEXP)?.[1];
+
+        if (gmailDelegatedAccountId) {
+          window.webContents.loadURL(url);
+
+          config.set(
+            "accounts",
+            config.get("accounts").map((account) => {
+              if (account.id === this.accountId) {
+                return {
+                  ...account,
+                  gmail: {
+                    ...account.gmail,
+                    delegatedAccountId: gmailDelegatedAccountId,
+                  },
+                };
+              }
+
+              return account;
+            }),
+          );
+
+          return {
+            action: "deny",
+          };
+        }
+
+        if (url === `${GMAIL_URL}/`) {
+          window.webContents.loadURL(url);
+
+          const account = accounts.getAccount(this.accountId);
+
+          if (account.config.gmail.delegatedAccountId) {
+            config.set(
+              "accounts",
+              config.get("accounts").map((account) => {
+                if (account.id === this.accountId) {
+                  return {
+                    ...account,
+                    gmail: {
+                      ...account.gmail,
+                      delegatedAccountId: null,
+                    },
+                  };
+                }
+
+                return account;
+              }),
+            );
+          }
+
+          return {
+            action: "deny",
+          };
+        }
+
+        const setupNewWindow = (newWindow: BrowserWindow) => {
+          this.registerNavigationHandler(newWindow);
+
+          this.registerWindowOpenHandler(newWindow);
+
+          setupWindowContextMenu(newWindow);
+
+          const account = accounts.getAccount(this.accountId);
+
+          account.instance.windows.add(newWindow);
+
+          const isUsingMultipleAccounts = accounts.getAccounts().length > 1;
+
+          if (config.get("googleApps.showAccountLabel") && isUsingMultipleAccounts) {
+            newWindow.webContents.on("page-title-updated", (_event, title) => {
+              newWindow.setTitle(`[${account.config.label}] ${title}`);
+            });
+          }
+
+          if (config.get("googleApps.showAccountColor") && isUsingMultipleAccounts) {
+            newWindow.webContents.on("dom-ready", () => {
+              if (account.config.color) {
+                const { value } = accountColorsMap[account.config.color];
+
+                ipc.renderer.send(
+                  newWindow.webContents,
+                  "googleApp.initAccountColorIndicator",
+                  value,
+                );
+              }
+            });
+          }
+
+          let powerSaveBlockerId: number | undefined;
+
+          if (matchedSupportedGoogleApp === "meet") {
+            powerSaveBlockerId = powerSaveBlocker.start("prevent-display-sleep");
+
+            globalShortcut.register("CommandOrControl+Shift+1", () => {
+              ipc.renderer.send(newWindow.webContents, "googleMeet.toggleMicrophone");
+            });
+
+            globalShortcut.register("CommandOrControl+Shift+2", () => {
+              ipc.renderer.send(newWindow.webContents, "googleMeet.toggleCamera");
+            });
+          }
+
+          newWindow.once("closed", () => {
+            account.instance.windows.delete(newWindow);
+
+            if (matchedSupportedGoogleApp === "meet") {
+              globalShortcut.unregister("CommandOrControl+Shift+1");
+              globalShortcut.unregister("CommandOrControl+Shift+2");
+            }
+
+            if (typeof powerSaveBlockerId === "number") {
+              powerSaveBlocker.stop(powerSaveBlockerId);
+            }
+          });
+        };
+
+        const newWindowOptions = {
+          autoHideMenuBar: true,
+          webPreferences: {
+            session: this.session,
+            preload: getPreloadPath("google-app"),
+          },
+        };
+
+        if (isGmailComposeWindowUrl(url)) {
+          return {
+            action: "allow",
+            createWindow: (inheritedOptions) => {
+              const newWindow = new BrowserWindow({
+                ...inheritedOptions,
+                ...newWindowOptions,
+                ...getCascadedWindowBounds({ width: 800, height: 600 }),
+                webPreferences: {
+                  ...inheritedOptions.webPreferences,
+                  ...newWindowOptions.webPreferences,
+                },
+              });
+
+              setupNewWindow(newWindow);
+
+              return newWindow.webContents;
+            },
+          };
+        }
+
+        const newGoogleAppWindow = new BrowserWindow({
+          ...newWindowOptions,
+          ...getCascadedWindowBounds({ width: 1280, height: 800 }),
+        });
+
+        setupNewWindow(newGoogleAppWindow);
+
+        newGoogleAppWindow.loadURL(url);
+
+        return {
+          action: "deny",
+        };
+      }
+
+      if (url.startsWith(`${GOOGLE_ACCOUNTS_URL}/AddSession`)) {
+        main.navigate("/settings/accounts");
+      } else if (WINDOW_OPEN_DOWNLOAD_URL_WHITELIST.some((regex) => regex.test(url))) {
+        window.webContents.downloadURL(url);
+      } else {
+        openExternalUrl(url, Boolean(matchedSupportedGoogleApp));
+      }
+
+      return {
+        action: "deny",
+      };
+    });
   }
 
   async fetchInboxFeed(fetchAttempt = 1) {
