--- packages/app/google-app-legacy.ts
@@ -1,477 +1,0 @@
-import { is } from "@electron-toolkit/utils";
-import { accountColorsMap } from "@meru/shared/accounts";
-import { APP_TITLEBAR_HEIGHT, GOOGLE_ACCOUNTS_URL } from "@meru/shared/constants";
-import {
-  GMAIL_DELEGATED_ACCOUNT_URL_REGEXP,
-  GMAIL_URL,
-  isGmailComposeWindowUrl,
-} from "@meru/shared/gmail";
-import { supportedGoogleApps } from "@meru/shared/types";
-import type { SupportedGoogleApp } from "@meru/shared/types";
-import {
-  BrowserWindow,
-  dialog,
-  globalShortcut,
-  powerSaveBlocker,
-  type Session,
-  WebContentsView,
-  type WebContentsViewConstructorOptions,
-} from "electron";
-import { createStore } from "zustand/vanilla";
-import { accounts } from "./accounts";
-import { config } from "./config";
-import { GoogleApp } from "./google-app";
-import { setupWindowContextMenu } from "./context-menu";
-import { ipc } from "./ipc";
-import { getCascadedWindowBounds, getPreloadPath } from "./lib/window";
-import { licenseKey } from "./license-key";
-import { main } from "./main";
-import { openExternalUrl } from "./url";
-import { subscribeWithSelector } from "zustand/middleware";
-
-const WINDOW_OPEN_URL_WHITELIST = [
-  /googleusercontent\.com\/viewer\/secure\/pdf/, // Print PDF
-];
-
-const SUPPORTED_GOOGLE_APPS_URL_REGEXP = new RegExp(
-  `(${Object.keys(supportedGoogleApps).join("|")})(?:\\.usercontent)?\\.google\\.com`,
-);
-
-const WINDOW_OPEN_DOWNLOAD_URL_WHITELIST = [/chat\.google\.com\/u\/\d\/api\/get_attachment_url/];
-
-type GoogleAppLegacyHooks = {
-  beforeLoadUrl?: ((view: WebContentsView) => void)[];
-};
-
-export type GoogleAppLegacyOptions = {
-  accountId: string;
-  url: string;
-  session: Session;
-  webContentsViewOptions?: WebContentsViewConstructorOptions;
-  hooks?: GoogleAppLegacyHooks;
-};
-
-export class GoogleAppLegacy {
-  accountId: string;
-
-  url: string;
-
-  baseUrl: string;
-
-  session: Session;
-
-  webContentsViewOptions: WebContentsViewConstructorOptions | undefined;
-
-  hooks: GoogleAppLegacyHooks = {};
-
-  private _view: WebContentsView | undefined;
-
-  get view() {
-    if (!this._view) {
-      throw new Error("View has not been created yet");
-    }
-
-    return this._view;
-  }
-
-  set view(view: WebContentsView) {
-    this._view = view;
-  }
-
-  viewStore = createStore(
-    subscribeWithSelector<{
-      navigationHistory: {
-        canGoBack: boolean;
-        canGoForward: boolean;
-      };
-      attentionRequired: boolean;
-    }>(() => ({
-      navigationHistory: {
-        canGoBack: false,
-        canGoForward: false,
-      },
-      attentionRequired: false,
-    })),
-  );
-
-  constructor({ accountId, url, session, webContentsViewOptions, hooks }: GoogleAppLegacyOptions) {
-    this.accountId = accountId;
-
-    this.url = url;
-
-    this.baseUrl = new URL(url).origin;
-
-    this.session = session;
-
-    this.webContentsViewOptions = webContentsViewOptions;
-
-    if (hooks) {
-      this.hooks = hooks;
-    }
-  }
-
-  createView(options?: WebContentsViewConstructorOptions) {
-    this.view = new WebContentsView({
-      ...this.webContentsViewOptions,
-      ...options,
-      webPreferences: {
-        ...this.webContentsViewOptions?.webPreferences,
-        ...options?.webPreferences,
-        session: this.session,
-      },
-    });
-
-    main.window.contentView.addChildView(this.view);
-
-    this.registerNavigationHandler(this.view);
-
-    this.registerFoundInPageHandler();
-
-    this.registerWindowOpenHandler(this.view);
-
-    this.view.webContents.on("dom-ready", () => {
-      this.view.webContents.setVisualZoomLevelLimits(1, 3);
-    });
-
-    setupWindowContextMenu(this.view);
-
-    this.updateViewBounds();
-
-    main.window.on("resize", () => {
-      this.updateViewBounds();
-    });
-
-    if (this.hooks.beforeLoadUrl) {
-      for (const hook of this.hooks.beforeLoadUrl) {
-        hook(this.view);
-      }
-    }
-
-    if (is.dev) {
-      this.view.webContents.openDevTools({ mode: "bottom" });
-    }
-
-    return this.view.webContents.loadURL(this.url);
-  }
-
-  private registerNavigationHandler(window: BrowserWindow | WebContentsView) {
-    window.webContents.on("did-navigate", (_event, url) => {
-      if (url.startsWith(`${GOOGLE_ACCOUNTS_URL}/v3/signin/challenge/pk/presend`)) {
-        dialog.showMessageBox({
-          type: "info",
-          message: "Passkey sign-in not supported yet",
-          detail: "Please use password to sign in.",
-        });
-      }
-
-      if (window === this.view) {
-        this.viewStore.setState({
-          navigationHistory: {
-            canGoBack: this.view.webContents.navigationHistory.canGoBack(),
-            canGoForward: this.view.webContents.navigationHistory.canGoForward(),
-          },
-          attentionRequired: !url.startsWith(this.baseUrl),
-        });
-      }
-    });
-
-    if (window === this.view) {
-      window.webContents.on("did-navigate-in-page", (_event: Electron.Event) => {
-        this.viewStore.setState({
-          navigationHistory: {
-            canGoBack: this.view.webContents.navigationHistory.canGoBack(),
-            canGoForward: this.view.webContents.navigationHistory.canGoForward(),
-          },
-        });
-      });
-    }
-
-    window.webContents.on("will-redirect", (event, url) => {
-      if (
-        url.startsWith("https://www.google.com") ||
-        url.startsWith("https://workspace.google.com")
-      ) {
-        event.preventDefault();
-
-        window.webContents.loadURL(`${GOOGLE_ACCOUNTS_URL}/ServiceLogin?service=mail`);
-      }
-    });
-  }
-
-  private registerFoundInPageHandler() {
-    this.view.webContents.on("found-in-page", (_event, result) => {
-      ipc.renderer.send(main.window.webContents, "findInPage.result", {
-        activeMatch: result.activeMatchOrdinal,
-        totalMatches: result.matches,
-      });
-    });
-  }
-
-  updateViewBounds() {
-    const { width, height } = main.getWindowBounds();
-
-    this.view.setBounds({
-      x: 0,
-      y: APP_TITLEBAR_HEIGHT,
-      width,
-      height: height - APP_TITLEBAR_HEIGHT,
-    });
-  }
-
-  destroy() {
-    this.view.webContents.removeAllListeners();
-
-    this.view.webContents.close();
-
-    this.view.removeAllListeners();
-
-    main.window.contentView.removeChildView(this.view);
-  }
-
-  registerWindowOpenHandler(window: BrowserWindow | WebContentsView) {
-    window.webContents.setWindowOpenHandler(({ url, disposition }) => {
-      if (url === "about:blank") {
-        return {
-          action: "allow",
-          createWindow: (options) => {
-            let newWindow: BrowserWindow | null = new BrowserWindow({
-              ...options,
-              show: false,
-            });
-
-            newWindow.webContents.once("will-navigate", (_event, url) => {
-              if (newWindow) {
-                if (url.startsWith(GOOGLE_ACCOUNTS_URL)) {
-                  newWindow.show();
-
-                  return;
-                }
-
-                openExternalUrl(url);
-
-                newWindow.webContents.close();
-
-                newWindow = null;
-              }
-            });
-
-            return newWindow.webContents;
-          },
-        };
-      }
-
-      if (
-        (url.startsWith(GMAIL_URL) && url.includes("view=pt")) ||
-        url.startsWith(GOOGLE_ACCOUNTS_URL)
-      ) {
-        return {
-          action: "allow",
-        };
-      }
-
-      const matchedSupportedGoogleApp = url.match(SUPPORTED_GOOGLE_APPS_URL_REGEXP)?.[1] as
-        | SupportedGoogleApp
-        | undefined;
-
-      const isGoogleAppEnabledToOpenInApp =
-        licenseKey.isValid &&
-        matchedSupportedGoogleApp &&
-        config.get("googleApps.openInApp") &&
-        !config.get("googleApps.openInAppExcludedApps").includes(matchedSupportedGoogleApp);
-
-      if (
-        (url.startsWith(GMAIL_URL) ||
-          WINDOW_OPEN_URL_WHITELIST.some((regex) => regex.test(url)) ||
-          isGoogleAppEnabledToOpenInApp) &&
-        disposition !== "background-tab"
-      ) {
-        if (matchedSupportedGoogleApp) {
-          if (
-            !config.get("googleApps.openAppsInNewWindow") &&
-            GoogleApp.reuseWindowByHostname(this.accountId, url)
-          ) {
-            return { action: "deny" };
-          }
-
-          new GoogleApp({
-            accountId: this.accountId,
-            url,
-          });
-
-          return { action: "deny" };
-        }
-
-        const gmailDelegatedAccountId = url.match(GMAIL_DELEGATED_ACCOUNT_URL_REGEXP)?.[1];
-
-        if (gmailDelegatedAccountId) {
-          window.webContents.loadURL(url);
-
-          config.set(
-            "accounts",
-            config.get("accounts").map((account) => {
-              if (account.id === this.accountId) {
-                return {
-                  ...account,
-                  gmail: {
-                    ...account.gmail,
-                    delegatedAccountId: gmailDelegatedAccountId,
-                  },
-                };
-              }
-
-              return account;
-            }),
-          );
-
-          return {
-            action: "deny",
-          };
-        }
-
-        if (url === `${GMAIL_URL}/`) {
-          window.webContents.loadURL(url);
-
-          const account = accounts.getAccount(this.accountId);
-
-          if (account.config.gmail.delegatedAccountId) {
-            config.set(
-              "accounts",
-              config.get("accounts").map((account) => {
-                if (account.id === this.accountId) {
-                  return {
-                    ...account,
-                    gmail: {
-                      ...account.gmail,
-                      delegatedAccountId: null,
-                    },
-                  };
-                }
-
-                return account;
-              }),
-            );
-          }
-
-          return {
-            action: "deny",
-          };
-        }
-
-        const setupNewWindow = (newWindow: BrowserWindow) => {
-          this.registerNavigationHandler(newWindow);
-
-          this.registerWindowOpenHandler(newWindow);
-
-          setupWindowContextMenu(newWindow);
-
-          const account = accounts.getAccount(this.accountId);
-
-          account.instance.windows.add(newWindow);
-
-          const isUsingMultipleAccounts = accounts.getAccounts().length > 1;
-
-          if (config.get("googleApps.showAccountLabel") && isUsingMultipleAccounts) {
-            newWindow.webContents.on("page-title-updated", (_event, title) => {
-              newWindow.setTitle(`[${account.config.label}] ${title}`);
-            });
-          }
-
-          if (config.get("googleApps.showAccountColor") && isUsingMultipleAccounts) {
-            newWindow.webContents.on("dom-ready", () => {
-              if (account.config.color) {
-                const { value } = accountColorsMap[account.config.color];
-
-                ipc.renderer.send(
-                  newWindow.webContents,
-                  "googleApp.initAccountColorIndicator",
-                  value,
-                );
-              }
-            });
-          }
-
-          let powerSaveBlockerId: number | undefined;
-
-          if (matchedSupportedGoogleApp === "meet") {
-            powerSaveBlockerId = powerSaveBlocker.start("prevent-display-sleep");
-
-            globalShortcut.register("CommandOrControl+Shift+1", () => {
-              ipc.renderer.send(newWindow.webContents, "googleMeet.toggleMicrophone");
-            });
-
-            globalShortcut.register("CommandOrControl+Shift+2", () => {
-              ipc.renderer.send(newWindow.webContents, "googleMeet.toggleCamera");
-            });
-          }
-
-          newWindow.once("closed", () => {
-            account.instance.windows.delete(newWindow);
-
-            if (matchedSupportedGoogleApp === "meet") {
-              globalShortcut.unregister("CommandOrControl+Shift+1");
-              globalShortcut.unregister("CommandOrControl+Shift+2");
-            }
-
-            if (typeof powerSaveBlockerId === "number") {
-              powerSaveBlocker.stop(powerSaveBlockerId);
-            }
-          });
-        };
-
-        const newWindowOptions = {
-          autoHideMenuBar: true,
-          webPreferences: {
-            session: this.session,
-            preload: getPreloadPath("google-app"),
-          },
-        };
-
-        if (isGmailComposeWindowUrl(url)) {
-          return {
-            action: "allow",
-            createWindow: (inheritedOptions) => {
-              const newWindow = new BrowserWindow({
-                ...inheritedOptions,
-                ...newWindowOptions,
-                ...getCascadedWindowBounds({ width: 800, height: 600 }),
-                webPreferences: {
-                  ...inheritedOptions.webPreferences,
-                  ...newWindowOptions.webPreferences,
-                },
-              });
-
-              setupNewWindow(newWindow);
-
-              return newWindow.webContents;
-            },
-          };
-        }
-
-        const newGoogleAppWindow = new BrowserWindow({
-          ...newWindowOptions,
-          ...getCascadedWindowBounds({ width: 1280, height: 800 }),
-        });
-
-        setupNewWindow(newGoogleAppWindow);
-
-        newGoogleAppWindow.loadURL(url);
-
-        return {
-          action: "deny",
-        };
-      }
-
-      if (url.startsWith(`${GOOGLE_ACCOUNTS_URL}/AddSession`)) {
-        main.navigate("/settings/accounts");
-      } else if (WINDOW_OPEN_DOWNLOAD_URL_WHITELIST.some((regex) => regex.test(url))) {
-        window.webContents.downloadURL(url);
-      } else {
-        openExternalUrl(url, Boolean(matchedSupportedGoogleApp));
-      }
-
-      return {
-        action: "deny",
-      };
-    });
-  }
-}
